updated delete character method

+++ Services/CharacterService/CharacterService.cs
@@ -31,10 +31,13 @@ namespace dotnet_rpg.Services.CharacterService
             var serviceResposne = new ServiceResponse<List<GetCharacterDto>>();
             
             Character character = _mapper.Map<Character>(newCharacter);                             //Since the input type was changed to AddCharacterDTO from Character... we now need to map the newCharacter to Character since they have different params
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
             //character.Id = characters.Max(c => c.Id) +1;                                            // returns the max characterID then adds 1
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
             serviceResposne.Data = await _context.Characters
+            .Where(c => c.User.Id == GetUserId())
             .Select(c => _mapper.Map<GetCharacterDto>(c))
             .ToListAsync();    //this is being done because the we still need to map the character into a GetCharDTO before we put it into the serviceResponse
             return serviceResposne;
@@ -42,24 +45,33 @@ namespace dotnet_rpg.Services.CharacterService
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> DeleteCharacter(int id)
         {
-            ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();         //QUESTION: the previous way didnt return the rest of the list for some reason so this needed to be changed
+            ServiceResponse<List<GetCharacterDto>> response = new ServiceResponse<List<GetCharacterDto>>();         //QUESTION: the previous way didnt return the rest of the list for some reason so this needed to be changed
 
 
             try
             {                                                                                        
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);                   //firstordefault returns null , first just returns an exception if not found
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-
-                serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                Character character = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());                   //firstordefault returns null , first just returns an exception if not found
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    response.Data = _context.Characters
+                        .Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else{
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
             }
             catch (Exception ex)                                                           
             {
-                serviceResponse.Success = false;
-                serviceResponse.Message = ex.Message;
+                response.Success = false;
+                response.Message = ex.Message;
             }
 
-            return serviceResponse;
+            return response;
         }
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
@@ -76,7 +88,8 @@ namespace dotnet_rpg.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter  = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter  = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);                     //this is being mapped... <class its being mapped to>(what is being mapped);
 
             return serviceResponse;
